// carrier/assets/gel-workspace/apps/company/src/types/configDetail/table.ts
import { TagProps } from '@wind/wind-ui/lib/tag/index'
import { ICfgDetailNodeCommonJson, IConfigDetailApiJSON, IConfigDetailTitleJSON } from './common.ts'
import { LinksModule, TLinkOptions } from '@/handle/link'
import { ReactNode } from 'react'
import { AxiosRequestConfig } from 'axios'

/**
 * TODO
 */
export type IConfigDetailTableColumnType =
  | 'array'
  // date 类型
  | 7
  // TODO
  | 4
  | 13
  | 12
  | 5
  // 货币
  | 6
  // 第二种 货币
  | 20
  | string
  | number
export type IConfigDetailTableColumnCustom =
  // 多个企业详情链接 name|id,name|id
  | 'CorpLinkArrStr'
  // 人物 合作企业
  | 'character_hzhb_1'
  // 招标产品
  | 'bidProduct'
  // 专利法律状态
  | 'patentLawStatus'
  | string

export type TConfigDetailTableColumn = {
  width: number | string
  dataIndex: string
  type?: IConfigDetailTableColumnType
  // 有 custom id type 会失效
  customId?: IConfigDetailTableColumnCustom
  /**
   * @default left
   */
  align?: 'left' | 'right' | 'center' | string | undefined

  // 是否是数组 , 如果是数组，对field每个值做相同处理
  isArray?: boolean
  links?: {
    // id的 key
    id: string
    titleKey?: string
    module: LinksModule
    standardLevelCodeKey?: string
    typeKey?: string
    standardTypeKey?: string // 标准信息的type key
  } & Pick<TLinkOptions, 'target'>
  info?: any

  tagInfo?: IConfigDetailTableTagInfo | IConfigDetailTableTagInfo[]
} & IConfigDetailTitleJSON

export interface IConfigDetailTableTagInfo {
  dataIndex: string
  color: string
  type?: TagProps['type']
  size?: TagProps['size']
}

export type ICfgDetailTableJson = {
  type: 'table' | 2 | undefined | string
  columns?: TConfigDetailTableColumn[] // 有可能使用父组件的 columns
  apiOptions?: AxiosRequestConfig | Record<string, any>
  maxTotal?: number
  dataSource?: any
  treeKey?: any
  expandable?: any
  params?: any
  hideTableHeader?: boolean
  titleRender?: (param: any) => ReactNode
  noExtra?: boolean
  titleRemark?: any
  nodes?: any
  searchOptions?: any[]
  nodeKey?: any
  showIndex?: boolean
  footerLeftRender?: any
  num?: any
} & IConfigDetailTitleJSON &
  IConfigDetailApiJSON &
  ICfgDetailNodeCommonJson